+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
 - Open groups detail panel on FCM group notification is received [#839](https://github.com/rokwire/illinois-app/issues/839).
+- Make the edit controls in phone login in Personal Info panel with white background to indicate that they are editable [#842](https://github.com/rokwire/illinois-app/issues/842).
 ### Fixed
 - Display all games in Athletics schedule [#857](https://github.com/rokwire/illinois-app/issues/857).